requirements.txt
Flask
gunicorn
google-genai
markdown
python-dotenv
groq
